Rename library
