+++ kadai2/rnakamine/lib/imgconv.go
